+++ .idea/encodings.xml
@@ -8,14 +8,14 @@
     <file url="file://$PROJECT_DIR$/client/src/main/resources" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/services/backstage-service/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/services/backstage-service/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/services/checkcard-service/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/services/checkcard-service/src/main/resources" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/services/demoservice-sayhelloworld/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/services/demoservice-sayhelloworld/src/main/resources" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/services/gateway-service/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/services/gateway-service/src/main/resources" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/services/grouputils-service/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/services/grouputils-service/src/main/resources" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/services/helloworld-service/src/main/java" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/services/helloworld-service/src/main/resources" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/services/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/services/src/main/resources" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/src/main/java" charset="UTF-8" />
