+++ services/checkcard-service/Dockerfile
@@ -1,0 +1,9 @@
+FROM amazoncorretto:8
+
+COPY target/checkcard-service-1.0.0-SNAPSHOT.jar /root/app.jar
+COPY start.sh /root/start.sh
+
+RUN chmod +x /root/start.sh
+
+ENTRYPOINT ["/root/start.sh"]
+EXPOSE 9901
