+++ services/checkcard-service/pom.xml
@@ -9,11 +9,11 @@
         <version>1.0.0-SNAPSHOT</version>
     </parent>
 
-    <artifactId>helloworld-service</artifactId>
+    <artifactId>checkcard-service</artifactId>
 
     <properties>
-        <maven.compiler.source>11</maven.compiler.source>
-        <maven.compiler.target>11</maven.compiler.target>
+        <maven.compiler.source>8</maven.compiler.source>
+        <maven.compiler.target>8</maven.compiler.target>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     </properties>
     <dependencies>
@@ -62,7 +62,7 @@
                 <artifactId>spring-boot-maven-plugin</artifactId>
                 <version>2.7.10</version>
                 <configuration>
-                    <mainClass>org.tongji.programming.DemoServiceProvider</mainClass>
+                    <mainClass>org.tongji.programming.CheckCardServiceProvider</mainClass>
                     <excludes>
                         <exclude>
                             <groupId>org.projectlombok</groupId>
