+++ services/checkcard-service/src/main/java/org/tongji/programming/config/CheckCardConfig.java
