+++ services/checkcard-service/src/main/java/org/tongji/programming/mapper/CheckResultMapper.java
