+++ services/checkcard-service/src/main/resources/Mapper/CheckResultMapper.xml
