+++ services/checkcard-service/start.sh
@@ -4,4 +4,4 @@
 #echo "172.17.0.1 host.docker.internal" >> /etc/hosts
 
 # Start the Java application
-java -Xmx512m -jar /root/app.jar
+java -Xmx400m -jar /root/app.jar
