+++ services/gateway-service/src/main/java/org/tongji/programming/controller/ChiakiController.java
@@ -1,13 +1,12 @@
 package org.tongji.programming.controller;
 
-import lombok.var;
 import org.apache.dubbo.config.annotation.DubboReference;
 import org.springframework.stereotype.Component;
 import org.tongji.programming.DTO.cqhttp.MessageUniversalReport;
 import org.tongji.programming.annotations.CommandMapper.*;
 import org.tongji.programming.enums.GroupLevel;
 import org.tongji.programming.service.DemoService;
-import org.tongji.programming.service.HelloWorldService;
+import org.tongji.programming.service.CheckCardService;
 
 @Component
 @CommandController("chiaki")
@@ -32,16 +31,16 @@ public class ChiakiController {
     }
 
     @DubboReference
-    HelloWorldService helloWorldService;
+    CheckCardService checkCardService;
 
     @CommandMapping("检查群名片")
     public String checkCard(MessageUniversalReport event){
-        return helloWorldService.checkCard(event.getGroupId(), event.getUserId());
+        return checkCardService.checkCard(event.getGroupId(), event.getUserId());
     }
 
     @CommandMapping("update")
     public String updateAssistants(MessageUniversalReport event){
-        return helloWorldService.addAssistants();
+        return checkCardService.addAssistants();
     }
